More robust workaround of Heroku read body
Read request body on all failures. This tries to ensures we respond
from the application instead of getting an Heroku error page because the
connection closed before sending a response.

This is a workaround for a limitation in Heroku who does not consider a
response from the application if the request body was not fully read.