"publish" page improvements
Created my first Hex package yesterday and felt this page could be
clearer.

The main improvement is mentioning the example mix.exs file early on in
the metadata section, since that's good to reference when reading the
description of the properties.

This update also aims to make the overview in the first sentence
clearer, and keeping its naming consistent with the headers below.

Also adds some cheer after successfully publishing the package, and
moves the technical details below the steps you must follow.