Move some release artifact handling out of transaction
On release delete, artifacts should only be removed and registry built
after transaction commits. If something breaks the registry will
eventually build again and artifacts can be removed when matching
against the database.

On publish, artifacts should be pushed inside the transaction, if the
transaction doesn't commit it doesn't matter if we pushed artifacts.
Registry should be built after commit because it will eventually build
and we do not want a registry that have releases that do not have
available artifacts or exists in database.