Add Bamboo.Formatter protocol for HexWeb.User
This protocol is used by Bamboo to format email addresses properly
for HexWeb.User structs. It allows passing in the `User` struct as
argument to the `to`, `from`, `cc` and `bcc` functions and formats
these e-mail fields in a more user-friendly fashion.

For more, check: https://hexdocs.pm/bamboo/Bamboo.Formatter.html