Move /keys endpoint under /repo namespace
This is somewhat of a hack because the Mix client stores a single API URL for the whole repository, so /api/repos/myrepo/keys would 404 if not under the /repo/:repository namespace.

The :repository is ignored so the hitting /api/keys is exactly the same as hitting /api/repos/myrepo/keys.