Simplify and limit scope of coc
The goal of this change is to remove redundancies in the text and focus
more on the positive rather than the negative. Instead of listing everything
that is not allowed we focus on communicating the goal of the code of
conduct and the spirit of the kind of communication and conduct we expect
from the community.

We are also limiting the scope of the coc to only govern the direct
community surrounding Hex and to areas where we can effectively enforce
it.