Add a blog RSS feed
My preferred way of keeping up with new posts on blogs is to subscribe
to their RSS feeds. I'd like to follow the Hex blog, but there was no
feed, so this commit/PR adds one.

There existed a feed for new packages around 2014/2015, from which I
took inspiration.

We reuse the `HexpmWeb.BlogView` view, since it already contains most of
what we need.

The feed is quite minimalistic with regards to the kinds of elements it
contains. Still, it is valid (with some recommendations) and serves its
purpose.

In order to support feed readers sorting posts by date, we include a
`pubDate` element on each post. This value is sourced from a new `time`
element added to each current post. It is parsed as a ISO 8601 timestamp
and then formatted as a ISO 2822 timestamp in order to follow the RSS
2.0 specification.

We implement our own `datetime_to_rfc2822` utility function. Pulling in
another dependency just for this didn't seem worth while.

Ref: <https://github.com/hexpm/hexpm/issues/854>
Ref: <https://github.com/hexpm/hexpm/pull/29>
Ref: <https://github.com/hexpm/hexpm/issues/19>
Ref: fba4d99765e441d99771f84f54385a99249d79fd
Ref: c96d46f400fbcfd702ed54a73e85cbb919e80834
Ref: <https://developer.mozilla.org/en-US/docs/Web/HTML/Element/time>
Ref: <https://validator.w3.org/feed/>