Add change state control pre and post show report
Interface displays controls for moderators only.
Only controls to change to an available state are displayed.
Appropiate message is displayed when there are no available states
to update the report to.

Report state change is validated on controller, as well as if
the action is beeing performed by a moderator.