Generate both write and read keys with a given name
Eric mentioned that for the hex client, both the write and read keys
are needed. This means that the previous flow does not exactly work,
as the user would have to do the flow twice.

I also noticed that the naming of keys was also not very good.

So, this commit:

- Changes the flow to generate both write and read keys
- Changes the flow to accept a user/client assigned name