Bypass controller and use context in helper functions
In the helper function `get_code` and `submit`, we used to use the
controller, while we could have just used the `WebAuth` context.

This is just redundant and inefficient as we have already have tests
for their respective functions.

This commit updates the helper functions and the test use the context
instead.