Allow getting release manifests without download count
Many build tools rely on checksums to ensure a remote resource didn't change unexpectedly,
Hex itself does it for package tarballs. Having the ever-changing `downloads` field in
the manifests makes it impossible to checksum them.

This change fixes that and allows the release manifests to be used in a hermetic manner.